Send_out_Translated/requirements.txt
sqlalchemy
psycopg2
flask
flask-cors-3.0.10
